pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.2.1",
    "faiss-cpu>=1.12.0",
    "fastapi[standard]>=0.120.0",
    "ipykernel>=7.0.1",
    "langchain>=1.0.2",
    "langchain-community>=0.4",
    "langchain-core>=1.0.0",
    "langchain-google-genai>=3.0.0",
    "langchain-groq>=1.0.0",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.1",
    "openai>=2.6.0",
    "pymupdf>=1.26.5",
    "pypdf>=6.1.3",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.2",
    "tqdm>=4.67.1",
]
